BlockwiseMergeFrontier.auto_fill(): fix the choice of `best_block`
Since the blocks always go back to the upper-left corner,
`block.get_original_indexes(0, 0)` is always referring to the same
merge, so the `max` wasn't doing anything. Actually, it was always
picking the first block, which is what we happen to want, but that is
because of the order that blocks are iterated over and implementation
details of `max`. So fix the logic so that the code works as intended.